requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0
langchain-core>=0.2.0
langchain-google-genai>=1.0.0
langgraph>=0.2.0
